@@ -1,13 +1,13 @@
 package net.rhizomik.rhizomer;
 
-import cucumber.api.CucumberOptions;
+import io.cucumber.junit.Cucumber;
+import io.cucumber.junit.CucumberOptions;
 import org.junit.runner.RunWith;
-import cucumber.api.junit.Cucumber;
 
 /**
  * Created by http://rhizomik.net/~roberto/
  */
 
 @RunWith(Cucumber.class)
-@CucumberOptions(plugin={"pretty"}, features="src/test/resources", tags = {"not @remote"})
+@CucumberOptions(plugin={"pretty"}, features="src/test/resources", tags = "not @remote")
 public class CucumberTest {}
