@@ -2,10 +2,12 @@ package net.rhizomik.rhizomer.model;
 
 import net.rhizomik.rhizomer.service.Queries;
 import net.rhizomik.rhizomer.service.OptimizedQueries;
+import org.hamcrest.CoreMatchers;
 import org.hamcrest.Matchers;
-import org.junit.Assert;
 import org.junit.Test;
 import org.springframework.util.StringUtils;
+
+import static org.hamcrest.MatcherAssert.*;
 
 /**
  * Created by http://rhizomik.net/~roberto/
@@ -17,58 +19,58 @@ public class QueriesTest {
     public void testNoAdditional2Samples() throws Exception {
         int classCount = 10, sampleSize = 5; double coverage = 0.5;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(0));
-        Assert.assertThat(selectsUnion, Matchers.is(""));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(0));
+        assertThat(selectsUnion, CoreMatchers.equalTo(""));
     }
 
     @Test
     public void testOneAdditional2Sample() throws Exception {
         int classCount = 10, sampleSize = 2; double coverage = 0.5;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(1));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 5 LIMIT 2"));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(1));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 5 LIMIT 2"));
     }
 
     @Test
     public void testTwoAdditional2Samples() throws Exception {
         int classCount = 10, sampleSize = 2; double coverage = 0.6;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(2));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 4 LIMIT 2"));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 8 LIMIT 2"));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(2));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 4 LIMIT 2"));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 8 LIMIT 2"));
     }
 
     @Test
     public void testNoAdditional1Samples() throws Exception {
         int classCount = 10, sampleSize = 1; double coverage = 0.1;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(0));
-        Assert.assertThat(selectsUnion, Matchers.is(""));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(0));
+        assertThat(selectsUnion, Matchers.is(""));
     }
 
     @Test
     public void testOneAdditional1Sample() throws Exception {
         int classCount = 10, sampleSize = 1; double coverage = 0.2;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(1));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 5 LIMIT 1"));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(1));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 5 LIMIT 1"));
     }
 
     @Test
     public void testThreeAdditional1Samples() throws Exception {
         int classCount = 10, sampleSize = 1; double coverage = 0.4;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(3));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 3 LIMIT 1"));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 6 LIMIT 1"));
-        Assert.assertThat(selectsUnion, Matchers.containsString("OFFSET 9 LIMIT 1"));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(3));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 3 LIMIT 1"));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 6 LIMIT 1"));
+        assertThat(selectsUnion, Matchers.containsString("OFFSET 9 LIMIT 1"));
     }
 
     @Test
     public void test11Samples() throws Exception {
         int classCount = 10, sampleSize = 11; double coverage = 0.5;
         String selectsUnion = queries.addSamples(classCount, sampleSize, coverage);
-        Assert.assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(0));
-        Assert.assertThat(selectsUnion, Matchers.is(""));
+        assertThat(StringUtils.countOccurrencesOf(selectsUnion, "\n"), Matchers.is(0));
+        assertThat(selectsUnion, Matchers.is(""));
     }
 }
